Refactor: R 사용 presentation -> designSystem 변경
