+++ designsystem/src/main/res/values/string.xml
@@ -1,0 +1,67 @@
+<resources>
+    <!-- group search -->
+    <string name="group_search_hint">모임 이름 검색</string>
+    <string name="group_search_register">우리 모임 새로 등록하기 +</string>
+    <string name="group_search_title">활동하고 계신\n보드게임 모임을 찾아주세요.</string>
+    <string name="group_search_list_description">지금 인기 있는 모임</string>
+    <string name="group_search_error_guide1">인터넷이 연결되어 있지 않습니다</string>
+    <string name="group_search_error_guide2">연결 상태를 다시 확인해주세요</string>
+    <string name="group_search_error_button">다시 불러오기</string>
+    <string name="group_search_not_found">‘%s’에 대한 검색 결과가 없습니다.</string>
+    <string name="group_search_not_found_guide">보드게임 모임을 등록해주세요!</string>
+    <string name="group_search_not_found_button">우리 보드게임 모임 만들러가기</string>
+
+    <string name="complete_message">모임 등록이 완료되었습니다!</string>
+    <string name="complete_group_organization_text">소속</string>
+    <string name="group_owner_text">관리자</string>
+    <string name="group_access_code_text">참여코드</string>
+    <string name="group_invite_guide">모임에 가입하기 위해서는 참여코드를 입력해야 \n합니다. 코드를 공유하고 멤버들을 초대해 보세요!</string>
+    <string name="confirmation_text">확인</string>
+    <string name="image_setting_delegate">모임 대표 이미지</string>
+    <string name="image_setting_delegate_guide">모임을 소개할 대표 이미지를 넣어보세요!</string>
+    <string name="image_setting_gallery">앨범에서 사진 선택</string>
+    <string name="image_setting_random">랜덤 이미지</string>
+    <string name="profile_setting">프로필 설정</string>
+    <string name="profile_setting_guide">관리자님이 모임에서 사용할 닉네임을 정해주세요</string>
+    <string name="profile_name_count">0/10</string>
+    <string name="complete_text">완료</string>
+
+    <string name="group_create_text">모임 등록</string>
+    <string name="group_name_text">보드게임 모임 이름</string>
+    <string name="group_name_hint">모임 이름을 입력해주세요.</string>
+    <string name="group_name_count">0/14</string>
+    <string name="group_description_text">모임 소개</string>
+    <string name="group_description_hint">모임을 소개해주세요.</string>
+    <string name="group_description_count">0/72</string>
+    <string name="group_organization_text">소속 (선택)</string>
+    <string name="group_organization_hint">소속을 알려주세요. (ex. 홍익대학교)</string>
+    <string name="group_organization_count">0/15</string>
+    <string name="group_create_guide">모임 등록하기</string>
+    <string name="group_creating_text">모임을 등록하는 중..</string>
+
+    <string name="game_result_record">게임 결과 기록</string>
+    <string name="game_select_guide">플레이 할 보드게임을 선택해주세요</string>
+    <string name="play_select_guide">게임을 함께 플레이 한 멤버를 선택해주세요</string>
+    <string name="player_complete_guide">플레이어 선택 완료</string>
+    <string name="member_search_hint">멤버 검색</string>
+
+    <string name="guest_add_text">임시 멤버 추가</string>
+    <string name="guest_description_text">아직 온보드에 가입하지 않은 상대를 모임에\n임시 멤버로 추가하고 게임 기록을 남겨보세요.</string>
+    <string name="guest_name_hint">임시 멤버가 사용할 닉네임을 입력해주세요.</string>
+    <string name="guest_name_set_tip">Tip :임시 멤버의 닉네임은 기억하기 쉽게 만들어주세요.\n이후 온보드에 가입할 때 기록을 연동할 수 있습니다.</string>
+
+    <string name="search_result_nothing">\'%s\'에 대한 검색 결과가 없습니다.</string>
+    <string name="search_result_nothing_guide">함께 보드게임을 플레이한 상대가 온보드에\n 아직 없다면,임시 멤버로 추가해 기록해보세요!</string>
+    <string name="guest_add_nothing">임시 멤버 추가</string>
+
+    <string name="game_result_rank">%d위</string>
+    <string name="game_player_count">%d인 플레이</string>
+    <string name="game_record_final_guide">최종 기록 이후 점수 수정이 불가능합니다\n정말로 확정하시겠습니까?</string>
+    <string name="game_record_guide">%s 결과를 입력해주세요!</string>
+    <string name="game_result_recording_text">점수 기록 중..</string>
+    <string name="game_result_recording_complete">%s\n점수 기록이 완료되었습니다.</string>
+    <string name="game_record_complete">기록 완료</string>
+    <string name="game_record_hint">0</string>
+
+    <string name="sign_up_name_hint">이름을 입력해주세요.</string>
+</resources>
