+++ presentation/src/main/java/com/yapp/bol/presentation/view/group/dialog/ProfileSettingDialog.kt
@@ -8,7 +8,7 @@ import android.os.Bundle
 import android.view.WindowManager
 import androidx.core.content.ContextCompat
 import androidx.core.widget.doOnTextChanged
-import com.yapp.bol.presentation.R
+import com.yapp.bol.designsystem.R
 import com.yapp.bol.presentation.databinding.ProfileSettingDialogBinding
 import com.yapp.bol.presentation.utils.Converter.convertLengthToString
 import com.yapp.bol.presentation.utils.dialogWidthResize
